move corpus.py to python module
