+++ backend/addcorpus/tests/mock_csv_corpus.py
@@ -1,4 +1,4 @@
-from addcorpus.corpus import CSVCorpusDefinition, FieldDefinition
+from addcorpus.python_corpora.corpus import CSVCorpusDefinition, FieldDefinition
 from addcorpus.extract import CSV
 import os
 import datetime
