+++ backend/corpora/dutchannualreports/dutchannualreports.py
@@ -9,7 +9,7 @@ from django.conf import settings
 
 from addcorpus.extract import XML, Metadata, Combined
 from addcorpus.filters import MultipleChoiceFilter, RangeFilter
-from addcorpus.corpus import XMLCorpusDefinition, FieldDefinition
+from addcorpus.python_corpora.corpus import XMLCorpusDefinition, FieldDefinition
 from media.image_processing import get_pdf_info, retrieve_pdf, pdf_pages, build_partial_pdf
 from addcorpus.python_corpora.load_corpus import corpus_dir
 from addcorpus.es_mappings import keyword_mapping, main_content_mapping
