+++ backend/corpora/hum_course_descriptions/hum_course_descriptions.py
@@ -5,7 +5,7 @@ import re
 from django.utils.html import strip_tags
 from langdetect import detect
 
-from addcorpus.corpus import FieldDefinition, XLSXCorpusDefinition
+from addcorpus.python_corpora.corpus import FieldDefinition, XLSXCorpusDefinition
 from addcorpus.es_mappings import text_mapping, main_content_mapping, keyword_mapping, int_mapping
 from addcorpus.extract import CSV, Combined, Pass, Constant, Metadata
 from addcorpus.filters import MultipleChoiceFilter
