+++ backend/corpora/parliament/norway-new.py
@@ -3,7 +3,7 @@ from datetime import datetime
 from django.conf import settings
 
 from addcorpus.extract import Combined, Constant, CSV
-from addcorpus.corpus import CSVCorpusDefinition
+from addcorpus.python_corpora.corpus import CSVCorpusDefinition
 from corpora.parliament.parliament import Parliament
 import corpora.parliament.utils.field_defaults as field_defaults
 import corpora.utils.formatting as formatting
