+++ backend/corpora/parliament/parliament.py
@@ -4,10 +4,9 @@ import os.path as op
 
 from django.conf import settings
 
-from addcorpus.corpus import CorpusDefinition
+from addcorpus.python_corpora.corpus import CorpusDefinition
 from addcorpus.filters import MultipleChoiceFilter
 import corpora.parliament.utils.field_defaults as field_defaults
-from corpora.parliament.utils.constants import MIN_DATE, MAX_DATE
 from addcorpus.es_settings import es_settings
 
 class Parliament(CorpusDefinition):
