+++ backend/corpora/parliament/sweden-old.py
@@ -1,7 +1,7 @@
 from glob import glob
 from datetime import datetime
 
-from addcorpus.corpus import CSVCorpusDefinition
+from addcorpus.python_corpora.corpus import CSVCorpusDefinition
 from addcorpus.extract import CSV, Constant
 from corpora.parliament.parliament import Parliament
 import corpora.parliament.utils.field_defaults as field_defaults
