+++ backend/corpora/peaceportal/peaceportal.py
@@ -5,7 +5,7 @@ from langdetect.lang_detect_exception import LangDetectException
 
 from django.conf import settings
 
-from addcorpus.corpus import ParentCorpusDefinition
+from addcorpus.python_corpora.corpus import ParentCorpusDefinition
 from addcorpus.extract import XML
 from addcorpus.es_settings import es_settings
 from corpora.peaceportal.utils import field_defaults
