+++ backend/corpora/utils/test_corpora_utils.py
@@ -1,4 +1,4 @@
-from addcorpus.corpus import FieldDefinition
+from addcorpus.python_corpora.corpus import FieldDefinition
 from addcorpus.extract import Constant
 
 from corpora.utils import exclude_fields
