+++ backend/download/tests/mock_corpora/multilingual_mock_corpus.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 import os
 
-from addcorpus.corpus import FieldDefinition, CSVCorpusDefinition
+from addcorpus.python_corpora.corpus import FieldDefinition, CSVCorpusDefinition
 from addcorpus.es_mappings import keyword_mapping, text_mapping
 from addcorpus.extract import CSV
