+++ backend/tag/tests/tag_mock_corpus.py
@@ -1,7 +1,7 @@
 import os
 import datetime
 
-from addcorpus.corpus import CSVCorpusDefinition, FieldDefinition
+from addcorpus.python_corpora.corpus import CSVCorpusDefinition, FieldDefinition
 from addcorpus.extract import CSV
 from addcorpus.es_mappings import keyword_mapping, main_content_mapping
