+++ backend/visualization/tests/mock_corpora/large_mock_corpus.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 import random
 
-from addcorpus.corpus import CorpusDefinition, FieldDefinition
+from addcorpus.python_corpora.corpus import CorpusDefinition, FieldDefinition
 from addcorpus.es_mappings import date_mapping, text_mapping
 
 TOTAL_DOCUMENTS = 11000
