+++ backend/wordmodels/tests/mock-corpus/mock_corpus.py
@@ -1,5 +1,5 @@
 import datetime
-from addcorpus.corpus import CorpusDefinition, FieldDefinition
+from addcorpus.python_corpora.corpus import CorpusDefinition, FieldDefinition
 
 from os.path import abspath, dirname, join
